backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-PyMongo
python-dotenv
